32 bit port, fixed MS VC++ issue in radix, fixed -j not default anymore

+++ src/arena.cpp
@@ -13,7 +13,7 @@ Arena::~Arena () {
 }
 
 void Arena::prepare (size_t bytes) {
-  LOG ("preparing 'to' space of arena with %ld bytes", (long) bytes);
+  LOG ("preparing 'to' space of arena with %zd bytes", bytes);
   assert (!to.start);
   to.top = to.start = new char[bytes];
   to.end = to.start + bytes;
@@ -21,8 +21,8 @@ void Arena::prepare (size_t bytes) {
 
 void Arena::swap () {
   delete [] from.start;
-  LOG ("delete 'from' space of arena with %ld bytes",
-    (long) (from.end - from.start));
+  LOG ("delete 'from' space of arena with %zd bytes",
+    (size_t) (from.end - from.start));
   from = to;
   to.start = to.top = to.end = 0;
 }
